LPS-131312 Rename to match the base cluster module
