+++ build-test.xml
@@ -6716,7 +6716,7 @@ go</echo>
 		<if>
 			<isset property="cluster.invoke.method" />
 			<then>
-				<echo file="${liferay.home}/osgi/configs/com.liferay.portal.cluster.sample.configuration.ClusterTestConfiguration.config">clusterTestCommand="${cluster.invoke.method}"</echo>
+				<echo file="${liferay.home}/osgi/configs/com.liferay.portal.cluster.multiple.sample.configuration.ClusterTestConfiguration.config">clusterTestCommand="${cluster.invoke.method}"</echo>
 			</then>
 		</if>
 	</target>
