+++ modules/apps/portal/portal-cluster-multiple-sample/src/main/java/com/liferay/portal/cluster/multiple/sample/internal/ClusterMethodInvoker.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cluster.sample.internal;
+package com.liferay.portal.cluster.multiple.sample.internal;
 
 import com.liferay.portal.kernel.cluster.ClusterExecutor;
 import com.liferay.portal.kernel.cluster.ClusterMasterExecutor;
