+++ modules/apps/portal/portal-cluster-multiple-sample/src/main/java/com/liferay/portal/cluster/multiple/sample/internal/SchedulerTestMessageListener.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cluster.sample.internal;
+package com.liferay.portal.cluster.multiple.sample.internal;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
