+++ modules/apps/portal/portal-cluster-multiple-sample/src/main/resources/META-INF/module-log4j.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE log4j:configuration SYSTEM "log4j.dtd">
 
 <log4j:configuration xmlns:log4j="http://jakarta.apache.org/log4j/">
-	<category name="com.liferay.portal.cluster.sample.internal">
+	<category name="com.liferay.portal.cluster.multiple.sample.internal">
 		<priority value="INFO" />
 	</category>
 </log4j:configuration>
