+++ portal-web/test/functional/com/liferay/portalweb/dependencies/groovy-script-portal-cache-put-modules.groovy
@@ -6,9 +6,9 @@ import com.liferay.portal.kernel.cache.PortalCacheManagerNames;
 import java.lang.reflect.Constructor;
 
 try {
-	ClassLoader testModuleClassLoader = ClassLoaderPool.getClassLoader("com.liferay.portal.cluster.sample_1.0.0");
+	ClassLoader testModuleClassLoader = ClassLoaderPool.getClassLoader("com.liferay.portal.cluster.multiple.sample_1.0.0");
 
-	Class<?> clazz = testModuleClassLoader.loadClass("com.liferay.portal.cluster.sample.internal.ClusterTestClass");
+	Class<?> clazz = testModuleClassLoader.loadClass("com.liferay.portal.cluster.multiple.sample.internal.ClusterTestClass");
 
 	Constructor<?> constructor = clazz.getConstructor(String.class);
