+++ portal-web/test/functional/com/liferay/portalweb/macros/Clustering.macro
@@ -27,7 +27,7 @@ definition {
 
 		AntCommand(
 			locator1 = "build-test.xml",
-			value1 = "deploy-osgi-module-to-specific-bundle -Dmodule.base.dir=modules/apps/portal/portal-cluster-sample -DliferayHome=${liferayHome}");
+			value1 = "deploy-osgi-module-to-specific-bundle -Dmodule.base.dir=modules/apps/portal/portal-cluster-multiple-sample -DliferayHome=${liferayHome}");
 	}
 
 	macro deployModuleOnServer {
